requirements.txt
detoxify==0.5.1
Flask==2.3.2
kafka_python==2.0.2
nltk==3.7
pymongo==4.3.3
selenium==4.10.0
